+++ internal/server/auth.go
@@ -58,7 +58,12 @@ func doCodeAuth(ctx Context, user, pass, passcode, value string) error {
 func getAuthToken(r *http.Request) string {
 	value := r.Header.Get(header.Authorization)
 	if value == "" {
-		return ""
+		// check for Authorization as a query parameter
+		// https://xyz/api/movies/uuid/location?Authorization=Bearer+token
+		value = r.URL.Query().Get(header.Authorization)
+		if value == "" {
+			return ""
+		}
 	}
 	result := strings.Split(value, " ")
 	var token string